requirements.txt
annotated-types==0.6.0
anyio==4.3.0
argcomplete==1.10.3
asgiref==3.8.1
azure-core==1.30.1
azure-identity==1.16.0
azure-storage-blob==12.19.1
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.8.2
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
coloredlogs==15.0.1
compressed_rtf==1.0.6
cryptography==42.0.5
Deprecated==1.2.14
distro==1.9.0
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
fastapi==0.110.2
filelock==3.13.4
flatbuffers==24.3.25
fsspec==2024.3.1
google-auth==2.29.0
googleapis-common-protos==1.63.0
grpcio==1.62.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
IMAPClient==2.1.0
importlib-metadata==7.0.0
importlib_resources==6.4.0
isodate==0.6.1
jsonpatch==1.33
jsonpointer==2.4
kubernetes==29.0.0
langchain-core==0.1.45
langchain-text-splitters==0.0.1
langsmith==0.1.49
lxml==5.2.1
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.1.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.17.3
openai==1.23.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.1
overrides==7.7.0
packaging==23.2
pdfminer.six==20191110
pillow==10.3.0
portalocker==2.8.2
posthog==3.5.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
PyJWT==2.8.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-magic==0.4.13
python-multipart==0.0.9
python-pptx==0.6.23
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
starlette==0.37.2
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.19.1
tqdm==4.66.2
typer==0.12.3
typing_extensions==4.11.0
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
xlrd==1.2.0
XlsxWriter==3.2.0
zipp==3.18.1
